fix(cascader): Search panel display error
